% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{class}
\name{AthenaConnection}
\alias{AthenaConnection}
\alias{AthenaConnection-class}
\alias{show,AthenaConnection-method}
\alias{dbQuoteString,AthenaConnection,character-method}
\alias{dbQuoteIdentifier,AthenaConnection,SQL-method}
\alias{dbListTables,AthenaConnection-method}
\alias{dbListTables}
\alias{dbListFields,AthenaConnection,character-method}
\alias{dbListFields}
\alias{dbExistsTable,AthenaConnection,character-method}
\alias{dbRemoveTable,AthenaConnection,character-method}
\alias{dbGetQuery,AthenaConnection,character-method}
\title{Athena Connection Methods}
\usage{
\S4method{show}{AthenaConnection}(object)

\S4method{dbQuoteString}{AthenaConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{AthenaConnection,SQL}(conn, x, ...)

\S4method{dbListTables}{AthenaConnection}(conn, ...)

\S4method{dbListFields}{AthenaConnection,character}(conn, name, ...)

\S4method{dbExistsTable}{AthenaConnection,character}(conn, name, ...)

\S4method{dbRemoveTable}{AthenaConnection,character}(conn, name, ...)

\S4method{dbGetQuery}{AthenaConnection,character}(conn, statement = NULL,
  ...)
}
\arguments{
\item{object}{Any R object}

\item{conn}{A subclass of \linkS4class{DBIConnection}, representing
an active connection to an DBMS.}

\item{x}{A character vector to quote as string.}

\item{...}{any other parameter for \code{Boto3} session: 
\href{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/core/session.html}{Boto3 session documentation}}

\item{name}{a character string with the name of the remote table.}

\item{statement}{a character string containing SQL.}
}
\description{
Implementations of pure virtual functions defined in the `DBI` package
for AthenaConnection objects.
}
\seealso{
\code{\link[DBI]{dbQuoteString}}, \code{\link[DBI]{dbQuoteIdentifier}}, \code{\link[DBI]{dbListTables}},
         \code{\link[DBI]{dbListFields}}, \code{\link[DBI]{dbExistsTable}}, \code{\link[DBI]{dbRemoveTable}},
         \code{\link[DBI]{dbGetQuery}}
}
