% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\docType{methods}
\name{AthenaTables}
\alias{AthenaTables}
\alias{sqlCreateTable,AthenaConnection-method}
\title{Athena Tables Methods}
\usage{
\S4method{sqlData}{AthenaConnection}(con, value, row.names = NA, ...)

\S4method{sqlCreateTable}{AthenaConnection}(con, table = NULL,
  fields = NULL, field.types = NULL, partition = NULL,
  s3.location = NULL, file.type = c("csv", "tsv", "parquet"), ...)
}
\arguments{
\item{con}{A database connection.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{...}{Other parameters passed on to methods.}

\item{table}{Name of the table. Escaped with
\code{\link[=dbQuoteIdentifier]{dbQuoteIdentifier()}}.}

\item{fields}{Either a character vector or a data frame.

A named character vector: Names are column names, values are types.
Names are escaped with \code{\link[=dbQuoteIdentifier]{dbQuoteIdentifier()}}.
Field types are unescaped.

A data frame: field types are generated using
\code{\link[=dbDataType]{dbDataType()}}.}

\item{field.types}{Additional field types used to override derived types.}
}
\description{
Implementations of pure virtual functions defined in the `DBI` package
for AthenaConnection objects
}
