% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\docType{methods}
\name{AthenaWriteTables}
\alias{AthenaWriteTables}
\alias{dbWriteTable,AthenaConnection,character,data.frame-method}
\alias{dbWriteTable,AthenaConnection,Id,data.frame-method}
\alias{dbWriteTable,AthenaConnection,SQL,data.frame-method}
\title{Convenience functions for reading/writing DBMS tables}
\usage{
\S4method{dbWriteTable}{AthenaConnection,character,data.frame}(conn, name,
  value, overwrite = FALSE, append = FALSE, row.names = NA,
  field.types = NULL, partition = NULL, s3.location,
  file.type = c("csv", "tsv", "parquet"), ...)

\S4method{dbWriteTable}{AthenaConnection,Id,data.frame}(conn, name, value,
  overwrite = FALSE, append = FALSE, row.names = NA,
  field.types = NULL, partition = NULL, s3.location,
  file.type = c("csv", "tsv", "parquet"), ...)

\S4method{dbWriteTable}{AthenaConnection,SQL,data.frame}(conn, name, value,
  overwrite = FALSE, append = FALSE, row.names = NA,
  field.types = NULL, partition = NULL, s3.location,
  file.type = c("csv", "tsv", "parquet"), ...)
}
\arguments{
\item{conn}{a \code{\linkS4class{AthenaConnection}} object, produced by
[DBI::dbConnect()]}

\item{name}{a character string specifying a table name. Names will be
automatically quoted so you can use any sequence of characters, not
just any valid bare table name.}

\item{value}{A data.frame to write to the database.}

\item{overwrite}{Allow overwriting the destination table. Cannot be
`TRUE` if `append` is also `TRUE`.}

\item{append}{Allow appending to the destination table. Cannot be
`TRUE` if `overwrite` is also `TRUE`.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

If \code{TRUE}, always translate row names to a column called "row_names".
If \code{FALSE}, never translate row names. If \code{NA}, translate
rownames only if they're a character vector.

A string is equivalent to \code{TRUE}, but allows you to override the
default name.

For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{partition}{partition Athena table (needs to be a named list or vector) for exmaple: \code{c(var1 = "2019-20-13")}}

\item{s3.location}{s3 bucket to store athena table}

\item{file.type}{What file type to be stored on s3 currently on support ["csv", "tsv", "parquet"]}

\item{...}{Other arguments used by individual methods.}
}
\description{
Convenience functions for reading/writing DBMS tables
}
\examples{
\dontrun{
library(DBI)
con <- dbConnect(RAthena::Athena(),s3_staging_dir = "s3://mybucket/athena_query/")
dbListTables(con)
dbWriteTable(con, "mtcars", mtcars, partition=c("TIMESTAMP" = format(Sys.Date() + 1, "\%Y\%m\%d")), s3.location = "s3://mybucket/data/")
dbReadTable(con, "mtcars")

dbListTables(con)
dbExistsTable(con, "mtcars")

dbDisconnect(con)
}
}
