% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_integration.R
\name{db_compute}
\alias{db_compute}
\alias{db_compute.AthenaConnection}
\title{S3 implementation of \code{db_compute} for Athena}
\usage{
db_compute.AthenaConnection(con, table, sql, ...)
}
\arguments{
\item{con}{A \code{\link{dbConnect}} object, as returned by \code{dbConnect()}}

\item{table}{Table name if left default RAthena will use default from 'dplyr''s \code{compute} function.}

\item{sql}{SQL code to be sent to the data}

\item{...}{passes \code{RAthena} table creation parameters: [\code{file.type},\code{s3.location},\code{partition}]}
}
\value{
db_compute returns table name
}
\description{
This is a backend function for dplyr's \code{compute} function. Users won't be required to access and run this function.
}
\seealso{
\code{\link{db_save_query}}
}
