% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\docType{methods}
\name{dbFetch}
\alias{dbFetch}
\alias{dbFetch,AthenaQuery-method}
\title{Fetch records from previously executed query}
\usage{
\S4method{dbFetch}{AthenaQuery}(res, n = -1, ...)
}
\arguments{
\item{res}{An object inheriting from \linkS4class{DBIResult}, created by
\code{\link[=dbSendQuery]{dbSendQuery()}}.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{...}{Other arguments passed on to methods.}
}
\value{
\code{dbFetch()} returns a data frame.
}
\description{
Currently returns the top n elements (rows) from result set or returns entire table from Athena.
}
\examples{
\dontrun{
library(DBI)

# Demo connection to athena using profile name 
con <- dbConnect(RAthena::athena(),
                 profile_name = "YOUR_PROFILE_NAME",
                 s3_staging_dir = "s3://path/to/query/bucket/")

res <- dbSendQuery(con, "show databases")
dbFetch(res)

# Check if connection if valid after closing connection
dbDisconnect(con)
}
}
\seealso{
\code{\link[DBI]{dbFetch}}
}
