% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R, R/Result.R
\docType{methods}
\name{dbGetInfo}
\alias{dbGetInfo}
\alias{dbGetInfo,AthenaConnection-method}
\alias{dbGetInfo,AthenaQuery-method}
\title{Get DBMS metadata}
\usage{
\S4method{dbGetInfo}{AthenaConnection}(dbObj, ...)

\S4method{dbGetInfo}{AthenaQuery}(dbObj, ...)
}
\arguments{
\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}

\item{...}{Other arguments to methods.}
}
\value{
a named list
}
\description{
Get DBMS metadata
}
\examples{
\dontrun{
library(DBI)

# Demo connection to athena using profile name 
con <- dbConnect(RAthena::athena(),
                 profile_name = "YOUR_PROFILE_NAME",
                 s3_staging_dir = "s3://path/to/query/bucket/")
                 
# Returns metadata from connnection object
dbGetInfo(con)

# Return metadata from Athena query object
res <- dbSendQuery(con, "show databases")
dbGetInfo(res)

# disconnect from Athena
dbDisconnect(con)
}
}
\seealso{
\code{\link[DBI]{dbGetInfo}}
}
