% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\docType{methods}
\name{dbHasCompleted}
\alias{dbHasCompleted}
\alias{dbHasCompleted,AthenaQuery-method}
\title{Completion status}
\usage{
\S4method{dbHasCompleted}{AthenaQuery}(res, ...)
}
\arguments{
\item{res}{An object inheriting from \linkS4class{DBIResult}.}

\item{...}{Other arguments passed on to methods.}
}
\value{
\code{dbHasCompleted()} returns a logical scalar. \code{TRUE} if the query has completed, \code{FALSE} otherwise.
}
\description{
This method returns if the query has completed.
}
\examples{
\dontrun{
# Note: 
# - Require AWS Account to run below example.
# - Different connection methods can be used please see `RAthena::dbConnect` documnentation

library(DBI)

# Demo connection to Athena using profile name 
con <- dbConnect(RAthena::athena(),
                 profile_name = "YOUR_PROFILE_NAME",
                 s3_staging_dir = "s3://path/to/query/bucket/")

# Check if query has completed
res <- dbSendQuery(con, "show databases")
dbHasCompleted(res)

# Check if connection if valid after closing connection
dbDisconnect(con)
}
}
\seealso{
\code{\link[DBI]{dbHasCompleted}}
}
