% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{methods}
\name{dbListFields}
\alias{dbListFields}
\title{List Field names of Athena table}
\usage{
\S4method{dbListFields}{AthenaConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{name}{a character string with the name of the remote table.}

\item{...}{Other parameters passed on to methods.}
}
\value{
\code{dbListFields()} returns a character vector with all the fields from an Athena table.
}
\description{
List Field names of Athena table
}
\examples{
\dontrun{
library(DBI)

# Connect to athena using default profile name
con <- dbConnect(RAthena::Athena(), s3_staging_dir = "s3://mybucket/athena_query/")

# Write data.frame to Athena table
dbWriteTable(con, "mtcars", mtcars,
             partition=c("TIMESTAMP" = format(Sys.Date(), "\%Y\%m\%d")),
             s3.location = "s3://mybucket/data/")
             
# Return list of fields in table
dbListFields(con, "mtcars")

# Disconnect conenction
dbDisconnect(con)
}
}
\seealso{
\code{\link[DBI]{dbListFields}}
}
