% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Connection.R
\docType{methods}
\name{dbListTables}
\alias{dbListTables}
\title{List Athena Tables}
\usage{
\S4method{dbListTables}{AthenaConnection}(conn, ...)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Other parameters passed on to methods.}
}
\value{
\code{dbListTables()} returns a character vector with all the tables from Athena.
}
\description{
Returns the unquoted names of Athena tables accessible through this connection.
}
\examples{
\dontrun{
library(DBI)

# Connect to athena using default profile name
con <- dbConnect(RAthena::Athena(), s3_staging_dir = "s3://mybucket/athena_query/")
             
# Return list of tables in Athena
dbListTables(con)

# Disconnect conenction
dbDisconnect(con)
}
}
\seealso{
\code{\link[DBI]{dbListTables}}
}
