% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAthena.R
\docType{package}
\name{RAthena-package}
\title{RAthena: Boto3 Athena Driver for DBI Backend}
\description{
Utilises the python boto3 sdk into athena to create a DBI driver.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.jones@shopdirect.com}

}
