% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Result.R
\docType{methods}
\name{AthenaResult}
\alias{AthenaResult}
\alias{AthenaQuery-class}
\alias{dbClearResult,AthenaQuery-method}
\alias{dbFetch,AthenaQuery-method}
\alias{dbHasCompleted,AthenaQuery-method}
\alias{dbGetInfo,AthenaQuery-method}
\title{Athena Result Methods}
\usage{
\S4method{dbClearResult}{AthenaQuery}(res, ...)

\S4method{dbFetch}{AthenaQuery}(res, n = -1, ...)

\S4method{dbHasCompleted}{AthenaQuery}(res, ...)

\S4method{dbGetInfo}{AthenaQuery}(dbObj, ...)
}
\arguments{
\item{res}{An object inheriting from \linkS4class{DBIResult}.}

\item{...}{Other arguments passed on to methods.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{dbObj}{An object inheriting from \linkS4class{DBIObject},
i.e. \linkS4class{DBIDriver}, \linkS4class{DBIConnection},
or a \linkS4class{DBIResult}}
}
\description{
Implementations of pure virtual functions defined in the `DBI` package
for AthenaResult objects.
}
