% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\docType{methods}
\name{dbConnect,AthenaDriver-method}
\alias{dbConnect,AthenaDriver-method}
\alias{dbConnect}
\title{Connect to Athena using python's sdk boto3}
\usage{
\S4method{dbConnect}{AthenaDriver}(drv, aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  database = "default", s3_staging_dir = NULL, region_name = NULL,
  botocore_session = NULL, profile_name = NULL, ...)
}
\arguments{
\item{drv}{an object that inherits from \linkS4class{DBIDriver},
or an existing \linkS4class{DBIConnection}
object (in order to clone an existing connection).}

\item{aws_access_key_id}{The Data Source Name.}

\item{aws_secret_access_key}{The Data Source Name.}

\item{aws_session_token}{The Data Source Name.}

\item{database}{The Data Source Name.}

\item{s3_staging_dir}{The Data Source Name.}

\item{region_name}{The Data Source Name.}

\item{botocore_session}{The Data Source Name.}

\item{profile_name}{The Data Source Name.}

\item{...}{authentication arguments needed by the DBMS instance; these
typically include \code{user}, \code{password}, \code{host}, \code{port}, \code{dbname}, etc.
For details see the appropriate \code{DBIDriver}.}
}
\description{
Connect to Athena using python's sdk boto3
}
