% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Driver.R
\docType{methods}
\name{dbConnect,AthenaDriver-method}
\alias{dbConnect,AthenaDriver-method}
\alias{dbConnect}
\title{Connect to Athena using python's sdk boto3}
\usage{
\S4method{dbConnect}{AthenaDriver}(drv, aws_access_key_id = NULL,
  aws_secret_access_key = NULL, aws_session_token = NULL,
  schema_name = "default", work_group = NULL, poll_interval = 1L,
  encryption_option = c("NULL", "SSE_S3", "SSE_KMS", "CSE_KMS"),
  kms_key = NULL, s3_staging_dir = NULL, region_name = NULL,
  botocore_session = NULL, profile_name = NULL, ...)
}
\arguments{
\item{drv}{an object that inherits from \linkS4class{DBIDriver},
or an existing \linkS4class{DBIConnection}
object (in order to clone an existing connection).}

\item{aws_access_key_id}{AWS access key ID}

\item{aws_secret_access_key}{AWS secret access key}

\item{aws_session_token}{AWS temporary session token}

\item{schema_name}{The schema_name to which the connection belongs}

\item{work_group}{The name of the \href{https://aws.amazon.com/about-aws/whats-new/2019/02/athena_workgroups/}{work group} to run Athena queries , Currently defaulted to \code{NULL}.}

\item{poll_interval}{Amount of time took when checking query execution status. Default set to \code{1L}.}

\item{encryption_option}{Athena encryption at rest \href{https://docs.aws.amazon.com/athena/latest/ug/encryption.html}{link}. 
Supported Amazon S3 Encryption Options ["NULL", "SSE_S3", "SSE_KMS", "CSE_KMS"]. Connection will default to NULL,
usually changing this option is not required.}

\item{kms_key}{\href{https://docs.aws.amazon.com/kms/latest/developerguide/overview.html}{AWS Key Management Service}, 
please refer to \href{https://docs.aws.amazon.com/kms/latest/developerguide/concepts.html}{link} for more information around the concept.}

\item{s3_staging_dir}{The location in Amazon S3 where your query results are stored, such as \code{s3://path/to/query/bucket/}}

\item{region_name}{Default region when creating new connections}

\item{botocore_session}{Use this Botocore session instead of creating a new default one.}

\item{profile_name}{The name of a profile to use. If not given, then the default profile is used.
To set profile name, the \href{https://aws.amazon.com/cli/}{AWS Command Line Interface} (AWS CLI) will need to be configured.
To configure AWS CLI please refer to: \href{https://docs.aws.amazon.com/cli/latest/userguide/cli-chap-configure.html}{Configuring the AWS CLI}.}

\item{...}{Any other parameter for Boto3 session: \href{https://boto3.amazonaws.com/v1/documentation/api/latest/reference/core/session.html}{Boto3 session documentation}}
}
\value{
\code{dbConnect()} returns a s4 class. This object is used to communicate with AWS Athena.
}
\description{
Connect to Athena using python's sdk boto3
}
\examples{
\dontrun{
# Connect to Athena using your aws access keys
 library(DBI)
 con <- dbConnect(RAthena:athena(),
                  aws_access_key_id='YOUR_ACCESS_KEY_ID', # 
                  aws_secret_access_key='YOUR_SECRET_ACCESS_KEY',
                  s3_staging_dir='s3://YOUR_S3_BUCKET/path/to/',
                  region_name='us-west-2')
 dbDisconnect(con)
 
# Connect to Athena using your profile name
# Profile name can be created by using AWS CLI
 con <- dbConnect(RAthena::athena(),
                  profile_name = "YOUR_PROFILE_NAME",
                  s3_staging_dir = "s3://path/to/query/bucket/")
 dbDisconnect(con)
}
}
\seealso{
\code{\link[DBI]{dbConnect}}
}
