% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAthena.R
\docType{package}
\name{RAthena-package}
\title{RAthena: Connect to Athena using Boto3 SDK (using the DBI interface)}
\description{
A DBI-compatible interface to Athena using python boto3 sdk.
}
\author{
\strong{Maintainer}: Dyfan Jones \email{dyfan.jones@shopdirect.com}

}
