% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Options.R
\name{RAthena_options}
\alias{RAthena_options}
\title{A method to change RAthena backend file parser.}
\usage{
RAthena_options(file_parser = c("data.table", "vroom"))
}
\arguments{
\item{file_parser}{Method to read and write tables to Athena, currently defaults to data.table}
}
\value{
\code{RAthena_options()} returns \code{NULL}, invisibly.
}
\description{
A method to change RAthena backend file parser.
}
\examples{
library(RAthena)

# change file parser from default data.table to vroom
RAthena_options("vroom")
}
